Grant MLA dataset upload access, raise upload cap to 10GB, fix clone hygiene
- deps.py/datasets.py: MLA's global read-only guard now exempts
  /api/v1/datasets/upload (and /api/v1/activity/log) so MLA can load
  geospatial datasets for review while staying blocked from every other
  mutation.
- init_db.py/session.py: schema (incl. new size_bytes BIGINT column) is
  now mirrored into every role-specific database on startup, so the
  role-routed get_db dependency doesn't hit missing-table/column errors
  for non-admin roles.
- config.py/dataset.py/nginx confs: raise MAX_UPLOAD_BYTES and the nginx
  client_max_body_size from 2GB to 10GB to fit full-resolution
  orthomosaics/DSMs; widen datasets.size_bytes to BigInteger to match.
- PublicPortalHeader: swap the outdated deputy commissioner photo/name
  for the current Commissioner, Dr. N. Mahantesh.
- Repo hygiene: untrack the runtime nginx.pid file, ignore stray
  .codex-*-run/ agent worktrees and *.pid so a real secret-bearing .env
  can never be swept up by a broad `git add`.

+++ backend/app/models/dataset.py
@@ -5,7 +5,7 @@ import enum
 import uuid
 from datetime import date, datetime
 
-from sqlalchemy import Date, Enum as SAEnum, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, Date, Enum as SAEnum, ForeignKey, String
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -48,7 +48,7 @@ class Dataset(Base):
         nullable=False,
     )
     storage_key: Mapped[str | None] = mapped_column(String(1024), nullable=True)
-    size_bytes: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    size_bytes: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
 
     status: Mapped[DatasetStatus] = mapped_column(
         SAEnum(DatasetStatus, name="dataset_status", native_enum=False, length=32),
